update get data by id

+++ models.py
@@ -8,7 +8,7 @@ class User(db.Model):
     name = db.Column(db.String(30), nullable=False)
     password= db.Column(db.String(120), nullable=False)
     email = db.Column(db.String(30), nullable=False, unique=True)
-    phone = db.Column(db.Integer, nullable=False, unique=True)
+    phone = db.Column(db.Integer(15), nullable=False, unique=True)
     address = db.Column(db.String(50), nullable=False)
     public_id = db.Column(db.String, nullable=False)
     is_admin = db.Column(db.Boolean, default=False)
